requirements.txt
Flask
requests
python-dotenv
wikipedia-api